requirements.txt
pillow==11.1.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
importlib_resources==6.5.2
matplotlib==3.10.0
numpy==2.2.2
scikit_image==0.25.2
scikit_learn==1.6.2
scipy==1.15.2
